Site updates
Added 404 page.
Changed Legalities page into an About page, containing the legalities, a fork button, an issues button, and a donate button.
Removed login modal's "Bugs and Requests" and Legalities buttons, replacing with a button leading to the About page.
Removed login modal's X button and replaced it with a ? button, leading to the About page.