Donator's Update
With this update, users can now donate and then receive the password which, when entered, will place a donator's icon next to their name. This does not override moderator, administrator, or creator status.
Donations are now made via Gumroad "products" which, when purchased, gives the user a text file containing the password needed to obtain the donator's icon.